Обновление Executor. Прием сообщений начинается сразу при создании объекта

+++ CompTech.Ict/src/CompTech.Ict.Sample/Executor.cs
@@ -7,7 +7,7 @@ using System.IO;
 using System.Diagnostics;
 using System.Threading.Tasks.Dataflow;
 
-namespace CompTech.Ict.Sample
+namespace example
 {
     public enum ExecutorStatus
     {
@@ -19,48 +19,58 @@ namespace CompTech.Ict.Sample
 
     public class Executor
     {
-        
+        public ExecutorStatus Status { get; private set; } = ExecutorStatus.Idle;
         private class InputParameters
         {
-            public string ScriptSource { get; set; }
+            public string ScriptSource { get; set; } 
             public string[] Parameters { get; set; }
             public Action<string[]> CallBack { get; set; }
         }
-
-        public ExecutorStatus Status { get; private set; } = ExecutorStatus.Idle;
         private Semaphore semaphore;
-        private BufferBlock<InputParameters> queue = new BufferBlock<InputParameters>();
-
+        BufferBlock<InputParameters> queue = new BufferBlock<InputParameters>();
         private string ArgumentsToFiles(string[] arguments, string scriptSource, string id)
         {
-            string pathTemp = Path.Combine(Path.GetTempPath(), id);
+            string pathTemp = Path.Combine(Path.GetTempPath(),id);
             Directory.CreateDirectory(pathTemp); // индивидуальная папка
             int count = arguments.Length;
             for (int i = 0; i < count; i++)
             {
-                string pathToArgument = Path.Combine(pathTemp, $"{i}input.txt");
+                string pathToArgument = Path.Combine(pathTemp, $"{i}input.txt"); 
                 File.WriteAllText(pathToArgument, arguments[i]);
             }
             File.WriteAllText(Path.Combine(pathTemp, "source.py"), scriptSource);
             return pathTemp;
         }
-
-        private string[] FilesToArguments(string path)
+        
+        private string[] FilesToOutputs(string path)
         {
+            
             string[] files = Directory.GetFiles(path, "*output.txt");
             string[] result = new string[files.Length];
-            for (int i = 0; i < files.Length; i++)
-                result[i] = File.ReadAllText(files[i]);
-            return result;
-        }
+            for( int i=0;i< files.Length;i++)
+            {
+                using (FileStream fs = File.OpenRead(files[i]))
+                {
+                    // преобразуем строку в байты
+                    byte[] array = new byte[fs.Length];
+                    // считываем данные
+                    fs.Read(array, 0, array.Length);
+                    // декодируем байты в строку
+                    result[i] = System.Text.Encoding.Default.GetString(array);
+                }
 
+            }
+            return result;
+       
+        }
         public Executor(int n)
         {
-            semaphore = new Semaphore(n, n);
             Status = ExecutorStatus.Idle;
+            semaphore = new Semaphore(n, n);
+            Loop();
         }
 
-        public void Add(string name, string[] parameters, Action<string[]> callBack) // по одному
+        public void Add(string name, string[] parameters, Action<string[]> callBack)
         {
             queue.Post(new InputParameters()
             {
@@ -70,52 +80,58 @@ namespace CompTech.Ict.Sample
             });
         }
 
-        public async Task Loop()
+        public async void Loop()
         {
             Status = ExecutorStatus.Running;
 
-            while (await queue.OutputAvailableAsync())
+            while (await queue.OutputAvailableAsync()) 
             {
-                InputParameters tmp = queue.Receive();
-
+                InputParameters tmp = queue.Receive(); 
+                
                 semaphore.WaitOne();
 
                 Action taskingOperation = () =>
                 {
-                    string id = Guid.NewGuid().ToString();
-                    var path = ArgumentsToFiles(tmp.Parameters, tmp.ScriptSource, id);
-                    OperationRun(path, tmp.CallBack);
-                    string[] result = FilesToArguments(path);
-                    Console.WriteLine(result[0]);
-                    tmp.CallBack?.Invoke(result);
-                    Directory.Delete(path, true);
+                    if (tmp.CallBack != null)
+                    {
+                        string id = Guid.NewGuid().ToString();
+                        var path = ArgumentsToFiles(tmp.Parameters, tmp.ScriptSource, id);
+                        OperationRun(path, tmp.CallBack);
+                        string[] result = FilesToOutputs(path);
+                        tmp.CallBack(result);
+                        Directory.Delete(path, true);
+                    }
                     semaphore.Release();
                 };
 #pragma warning disable CS4014 // Так как этот вызов не ожидается, выполнение существующего метода продолжается до завершения вызова
-                Task.Run(taskingOperation);
+               Task.Run(taskingOperation);
 #pragma warning restore CS4014 // Так как этот вызов не ожидается, выполнение существующего метода продолжается до завершения вызова
             }
         }
 
         private void OperationRun(string inputPath, Action<string[]> callBack)
         {
-            ProcessStartInfo start = new ProcessStartInfo("python");
-            start.Arguments = inputPath + "\\source.py " + inputPath;
-            start.UseShellExecute = false;
-            start.RedirectStandardOutput = true;
-            string tmpResult = "";
+            ProcessStartInfo start = new ProcessStartInfo()
+            {
+                FileName = "python",
+                Arguments = Path.Combine(inputPath, "source.py") + " " + inputPath,
+                UseShellExecute = false,
+                RedirectStandardOutput = true
+            };        
+
             try
             {
                 Process process = Process.Start(start);
-                using (StreamReader reader = process.StandardOutput)
-                    tmpResult = reader.ReadToEnd();
+                process.WaitForExit();
+                if (process.ExitCode != 0)
+                    throw new Exception("Ошибка при выполнении скрипта");
             }
-            catch (Exception ex)
+            catch(Exception ex)
             {
+                Console.WriteLine(ex.Message);
                 callBack(null);
+                //Console.Error;
             }
         }
     }
 }
-
-
